notif_server/requirements.txt
blinker==1.9.0
click==8.1.7
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pymongo==4.10.1
python-dotenv==1.0.1
tenacity==9.0.0
Werkzeug==3.1.3
zipp==3.21.0
